model/requirements.txt
langchain>=0.1.20
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
pgvector>=0.2.5
psycopg2-binary>=2.9.9
PyPDF2>=3.0.1
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.0
sentence-transformers>=2.3.0
scikit-learn>=1.4.0
psycopg2-binary
python-docx 
pdfplumber 
openpyxl
google-generativeai
pymupdf